requirements.txt
# core app dependencies
streamlit

# google calendar/sheets
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
httplib2

# data handling
pandas
numpy
pytz
python-dateutil

# LLM/embeddings
llama_cpp_python
sentence-transformers
transformers
scikit-learn

# FAISS for similarity search
faiss-cpu
